% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actual_use_codes}
\alias{actual_use_codes}
\title{BC Assessment Manual Actual Use lookup table}
\format{
An tibble with the BCA actual use codes and descriptions.
}
\source{
Derived from the dataset
}
\usage{
data(actual_use_codes)
}
\description{
BC Assessment Manual Actual Use lookup table
}
\examples{
data(actual_use_codes)
}
\keyword{datasets}
