% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_bca_data}
\alias{get_bca_data}
\title{Get a connection to the BCA data table}
\usage{
get_bca_data(
  table,
  api_key = Sys.getenv("ABACUS_API_TOKEN"),
  version = list(revision = 21, persistent_id = "hdl:11272.1/AB2/LAPUAB/7ZYI7W"),
  cache_path = Sys.getenv("ABACUS_CACHE_PATH"),
  refresh = FALSE
)
}
\arguments{
\item{table}{name of the data table. Possibly values are
"address", "assessmentAreaKey", "commercialInventory", "folio", "folioDescription",
"folioDescriptionTax", "jurisdictionKey", "legalDescription" , "metadata",
"residentialInventory", "sales", "valuation", "value"}

\item{api_key}{Abacus API key}

\item{version}{database version, requires revision number and persistent ID}

\item{cache_path}{cache path for local caching of the database}

\item{refresh}{optionally refresh the local database if set to \code{TRUE}}
}
\value{
A database connection to the residential inventory table
}
\description{
Will download the data if necessary.
}
