% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_bca_inventory}
\alias{get_bca_inventory}
\title{Get inventory data}
\usage{
get_bca_inventory(
  version = list(revision = "24", series = c("residential_inventory",
    "commercial_inventory")),
  api_key = Sys.getenv("ABACUS_API_TOKEN"),
  cache_path = Sys.getenv("ABACUS_CACHE_PATH"),
  assessment_areas = NULL,
  refresh = FALSE
)
}
\arguments{
\item{version}{database version, requires revision number and persistent ID}

\item{api_key}{Abacus API key}

\item{cache_path}{cache path for local caching of the database}

\item{assessment_areas}{optionally filter the data by assessment areas}

\item{refresh}{optionally refresh the local database if set to \code{TRUE}}
}
\value{
A database connection to the residential inventory table
}
\description{
Will download the data if necessary.
}
