% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folios_spatial.R
\name{get_bca_spatial_folios}
\alias{get_bca_spatial_folios}
\title{Load BCA spatial polygons
#'}
\usage{
get_bca_spatial_folios(
  query = "select * from \\"WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_DESCRIPTIONS_SV",
  version = list(date = "20221005", persistent_id = "hdl:11272.1/AB2/NXRVP9/KTMVPW"),
  api_key = Sys.getenv("ABACUS_API_TOKEN"),
  cache_path = Sys.getenv("ABACUS_CACHE_PATH"),
  refresh = FALSE
)
}
\arguments{
\item{query}{query to select and filer layers. Available layers are
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_ADDRESSES_SV
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_LEGAL_DESCRIPTS_SV
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_GNRL_PROP_VALUES_SV
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_SALES_SV
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_DESCRIPTIONS_SV
* WHSE_HUMAN_CULTURAL_ECONOMIC_BCA_FOLIO_LAND_CHARS_SV}

\item{version}{database version, requires revision number and persistent ID}

\item{api_key}{Abacus API key}

\item{cache_path}{cache path for local caching of the database}

\item{refresh}{optionally refresh the local database if set to \code{TRUE}}
}
\value{
A database connection to the residential inventory table
}
\description{
Will download the data if necessary.
}
