% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folios_spatial.R
\name{get_bca_spatial_gpkg}
\alias{get_bca_spatial_gpkg}
\title{Load BCA spatial polygons
#'}
\usage{
get_bca_spatial_gpkg(
  version = list(revision = "24"),
  api_key = Sys.getenv("ABACUS_API_TOKEN"),
  cache_path = Sys.getenv("ABACUS_CACHE_PATH"),
  refresh = FALSE
)
}
\arguments{
\item{version}{database version, requires revision number and persistent ID}

\item{api_key}{Abacus API key}

\item{cache_path}{cache path for local caching of the database}

\item{refresh}{optionally refresh the local database if set to \code{TRUE}}
}
\value{
A database connection to the residential inventory table
}
\description{
Will download the data if necessary.
}
