% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual_class_codes}
\alias{manual_class_codes}
\title{BC Assessment Manual Class lookup table}
\format{
An tibble with the BCA manual class codes and descriptions.
}
\source{
\href{https://www.bcassessment.ca/Files/Misc/UserGuide/Tables/dataadv12.pdf}{BCA}
}
\usage{
data(manual_class_codes)
}
\description{
BC Assessment Manual Class lookup table
}
\examples{
data(manual_class_codes)
}
\keyword{datasets}
