% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{get_bca_sqlite_connection}
\alias{get_bca_sqlite_connection}
\title{Get connection to BCA database}
\usage{
get_bca_sqlite_connection(
  version = list(revision = 22, persistent_id = "hdl:11272.1/AB2/LAPUAB/HHU1YP"),
  api_key = Sys.getenv("ABACUS_API_TOKEN"),
  cache_path = Sys.getenv("ABACUS_CACHE_PATH"),
  refresh = FALSE
)
}
\arguments{
\item{version}{database version, requires revision number and persistent ID}

\item{api_key}{Abacus API key}

\item{cache_path}{cache path for local caching of the database}

\item{refresh}{optionally refresh the local database if set to \code{TRUE}}
}
\value{
A database connection
}
\description{
Will download the data if necessary.
}
